+++ src/styles.css
@@ -41,12 +41,14 @@ main {
   grid-template-rows: repeat(4, 1fr);
   row-gap: 10px;
   column-gap: 10px;
-  margin: 10px;
+  margin: 10px ; 
+ 
+  
 }
 
 .box-pagina-principal {
   display: grid;
-  grid-template-rows: 3fr 1fr;
+  grid-template-rows: 2fr ;
   grid-template-columns: 1fr;
   border-style: solid;
   border-width: thin;
@@ -62,13 +64,24 @@ footer {
   width: 100%;
   display: flex;
   padding: 0 10px;
+  height: 7%;
 }
 
 .box-pagina-principal > img {
   width: 100%;
-  height: 200px;
+  height: 150px;
 }
 
 h4 {
   margin-left: 20px;
+}
+
+.infoUsuario {
+  display: flex;
+  flex-direction: row;
+  gap: 5px;
+}
+.imgCard{
+  height: 40px;
+  border-radius: 50%;
 }
